refactor: detekt format
