+++ app/src/main/java/com/jonas/cookly/core/util/ServiceResult.kt
