+++ app/src/main/java/com/jonas/cookly/core/util/logging/LogFunctions.kt
