+++ app/src/main/java/com/jonas/cookly/core/util/sideeffect/SingleEventEffect.kt
