+++ app/src/main/java/com/jonas/cookly/ui/presentation/features/register/domain/model/RegisterInputValidationType.kt
