+++ app/src/main/java/com/jonas/cookly/ui/presentation/features/register/domain/source/RegisterUserRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.jonas.cookly.ui.presentation.features.register.domain.source
 
 import com.jonas.cookly.core.domain.SimplesResponseModel
-import com.jonas.cookly.ui.presentation.features.register.domain.model.AddUserRequestModel
 import com.jonas.cookly.core.util.ServiceResult
+import com.jonas.cookly.ui.presentation.features.register.domain.model.AddUserRequestModel
 
 interface RegisterUserRemoteDataSource {
     suspend fun registerUser(addUserRequestModel: AddUserRequestModel): ServiceResult<SimplesResponseModel>
