+++ app/src/main/java/com/jonas/cookly/ui/presentation/navigation/graphs/HomeGraph.kt
@@ -2,12 +2,8 @@ package com.jonas.cookly.ui.presentation.navigation.graphs
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
-import com.jonas.cookly.ui.presentation.navigation.screens.Graphs
-import com.jonas.cookly.ui.presentation.navigation.screens.HomeScreens
 
-fun NavGraphBuilder.homeGraph(
-    onNavigateUp: () -> Unit
-) {
+fun NavGraphBuilder.homeGraph() {
 //    navigation<Graphs.HomeGraph>(
 //        startDestination = HomeScreens.HomeScreen
 //    ) {
