+++ app/src/main/java/com/jonas/cookly/ui/presentation/navigation/screens/Screens.kt
@@ -50,5 +50,4 @@ sealed interface HomeScreens {
 
     @Serializable
     data object SearchScreen : HomeScreens
-
 }
