+++ app/src/main/java/com/jonas/cookly/ui/theme/Color.kt
@@ -36,4 +36,3 @@ val DarkSurface = Color(0xFF303030)
 val DarkOnSurface = Color(0xFFFFFFFF)
 
 val DarkText = Color(0xFFFFFFFF)
-
